requirements.txt
gradio>=4.29.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.2
sqlalchemy>=2.0.0
alembic>=1.12.0
huggingface_hub>=0.20.0
transformers>=4.38.0
torch>=2.1.0
langchain>=0.1.11
langgraph>=0.0.20
orjson>=3.9.10
pydantic>=2.6.0
python-dateutil>=2.8.2
joblib>=1.3.0
